# openapi.yaml
openapi: 3.0.1
info:
  title: Second checkpoint
  description: 'Second checkpoint'
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:8080/api/v1/
tags:
- name: item
  description: Items
  externalDocs:
    description: Find out more
    url: http://swagger.io
paths:
  /item/add/:
    post:
      tags:
      - item
      summary: Add a new item to the store
      operationId: addItem
      requestBody:
        description: Item object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemRequest'
        required: true
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        400:
          description: Bad request
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "SKU"
      security:
      - item_auth:
        - write:items
        - read:items
      x-codegen-request-body-name: body
  /item/{id}:
    get:
      tags:
      - item
      summary: Find item by ID
      description: Returns a single item
      operationId: getItemById
      parameters:
      - name: id
        in: path
        description: ID of item to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemResponse'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Item not found
          content: {}
      security:
      - item_auth:
        - write:items
        - read:items
components:
  schemas:
    ItemRequest:
      required:
      - sku
      - name
      - price
      - currency
      - thumbnail
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
          example: soap
        price:
          type: number
          format: double
        currency:
          type: string
        thumbnail:
          type: string
          format: uri
    ItemResponse:
      required:
      - sku
      - name
      - price
      - currency
      - thumbnail
      type: object
      properties:
        sku:
          type: string
        name:
          type: string
          example: soap
        price:
          type: number
          format: double
        currency:
          type: string
        thumbnail:
          type: string
          format: uri
    ApiResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
          example: 1
        message:
          type: string
          example: "success"
  securitySchemes:
    item_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://petstore.swagger.io/oauth/dialog
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets